refactored store code
