+++ sfaira/data/__init__.py
@@ -1,7 +1,10 @@
-from sfaira.data.base import clean_string, DatasetBase, \
+from sfaira.data.dataloaders.base import clean_string, DatasetBase, \
     DatasetGroup, DatasetGroupDirectoryOriented, \
-    DatasetSuperGroup, load_store, DistributedStoreBase, DistributedStoreH5ad, DistributedStoreDao
+    DatasetSuperGroup
+from sfaira.data.store import load_store, DistributedStoreMultipleFeatureSpaceBase, DistributedStoresH5ad, \
+    DistributedStoresDao
 from . import dataloaders
 from .dataloaders import Universe
 from .interactive import DatasetInteractive
+from . import store
 from . import utils
