+++ sfaira/data/dataloaders/__init__.py
@@ -1,3 +1,4 @@
+from . import base
 from . import databases
 from . import loaders
 from .super_group import Universe
