+++ sfaira/data/store/__init__.py
@@ -1,0 +1,2 @@
+from sfaira.data.store.multi_store import load_store, DistributedStoreMultipleFeatureSpaceBase, \
+    DistributedStoresH5ad, DistributedStoresDao
