+++ sfaira/train/train_model.py
@@ -6,19 +6,19 @@ import pickle
 from typing import Union
 
 from sfaira.consts import AdataIdsSfaira
-from sfaira.data import DistributedStoreBase, Universe
+from sfaira.data import DistributedStoreMultipleFeatureSpaceBase, Universe
 from sfaira.estimators import EstimatorKeras, EstimatorKerasCelltype, EstimatorKerasEmbedding
 from sfaira.ui import ModelZoo
 
 
 class TrainModel:
 
-    data: Union[anndata.AnnData, DistributedStoreBase]
+    data: Union[anndata.AnnData, DistributedStoreMultipleFeatureSpaceBase]
     estimator: Union[EstimatorKeras, None]
 
     def __init__(
             self,
-            data: Union[str, anndata.AnnData, Universe, DistributedStoreBase],
+            data: Union[str, anndata.AnnData, Universe, DistributedStoreMultipleFeatureSpaceBase],
     ):
         # Check if handling backed anndata or base path to directory of raw files:
         if isinstance(data, str) and data.split(".")[-1] == "h5ad":
@@ -30,7 +30,7 @@ class TrainModel:
             self.data = data
         elif isinstance(data, Universe):
             self.data = data.adata
-        elif isinstance(data, DistributedStoreBase):
+        elif isinstance(data, DistributedStoreMultipleFeatureSpaceBase):
             self.data = data
         else:
             raise ValueError(f"did not recongize data of type {type(data)}")
@@ -42,7 +42,7 @@ class TrainModel:
         Loads backed objects from DistributedStoreBase into single adata object in memory in .data slot.
         :return:
         """
-        if isinstance(self.data, DistributedStoreBase):
+        if isinstance(self.data, DistributedStoreMultipleFeatureSpaceBase):
             adata = None
             for k, v in self.data.indices.items():
                 x = self.data.adata_by_key[k][v, :].to_memory()
@@ -105,7 +105,7 @@ class TrainModelEmbedding(TrainModel):
     def __init__(
             self,
             model_path: str,
-            data: Union[str, anndata.AnnData, Universe, DistributedStoreBase],
+            data: Union[str, anndata.AnnData, Universe, DistributedStoreMultipleFeatureSpaceBase],
     ):
         super(TrainModelEmbedding, self).__init__(data=data)
         self.estimator = None
@@ -168,7 +168,7 @@ class TrainModelCelltype(TrainModel):
     def __init__(
             self,
             model_path: str,
-            data: Union[str, anndata.AnnData, Universe, DistributedStoreBase],
+            data: Union[str, anndata.AnnData, Universe, DistributedStoreMultipleFeatureSpaceBase],
             fn_target_universe: str,
     ):
         super(TrainModelCelltype, self).__init__(data=data)
